pull

+++ app/Http/Controllers/CityController.php
@@ -15,41 +15,41 @@ class CityController extends Controller
 
    	public function create()
    	{
-   	return view('city.create_city');
+   	    return view('city.create_city');
    	}
 
     public function store(Request $request) 
     {
-    $this->validate($request, [
-        'city_name' => 'required|unique:cities',
-    
-    ]);
-    $input = $request->all();
-    $input['created_by'] = " 1 ";
-    City::create($input);
-    $city=City::all();
-    return view('city.show_city',compact('city'));
+        $this->validate($request, [
+            'city_name' => 'required|unique:cities',
+        
+        ]);
+        $input = $request->all();
+        $input['created_by'] = " 1 ";
+        City::create($input);
+        $city=City::all();
+        return view('city.show_city',compact('city'));
     }
 
 
     public function delete($id)
     {
-    /*if(Venue::where('venue_city',$id)->first())
-    {
-        return("not possible its been used in venue ");
-    }*/
+        if(Venue::where('venue_city',$id)->first())
+        {
+            return("not possible its been used in venue ");
+        }
 
-   	City::where('city_id',$id)->delete();
-    $city=City::all();
-    return view('city.show_city',compact('city'));
+       	City::where('city_id',$id)->delete();
+        $city=City::all();
+        return view('city.show_city',compact('city'));
     
     }
 
 
     public function show()
     {
-    $city=City::all();
-    return view('city.show_city',compact('city'));
+        $city=City::all();
+        return view('city.show_city',compact('city'));
     }
 
     public function edit($id)
